Refactor dataGenerator/seed to run independent of the server
