--- helpers/index.js
@@ -1,5 +1,0 @@
-const dataGenerator = require('./dataGenerator');
-
-module.exports = {
-  dataGenerator
-};
