+++ package.json
@@ -7,7 +7,8 @@
   "author": "Steve McLaughlin <stevejrmc@gmail.com>",
   "license": "MIT",
   "scripts": {
-    "start": "nodemon server/app.js"
+    "start": "nodemon server/app.js",
+    "seed": "node helpers/seed.js"
   },
   "dependencies": {
     "aws-sdk": "^2.585.0",
